Fix .fromConnectionString test
After the commit
https://github.com/Azure/amqp-common-js/commit/a2a8e7d5b242d278dbe9e198d07194c32e3d24cf#diff-ad0ce2742650d03b78af445d57e30269
our test now encounters a different error while invoking `.endsWith()` on
`parsedCS.Endpoint` of `undefined` at
https://github.com/Azure/amqp-common-js/commit/a2a8e7d5b242d278dbe9e198d07194c32e3d24cf#diff-ad0ce2742650d03b78af445d57e30269L74.
(The issue of accessing undefined endpoint no longer exists in the latest
amqp-common code.)

The better place to test this error is in the amqp-common-js repo because it is
thrown by the `EventHubConnectionConfig.create()` method. This change removes
the check in `EventHubClient.createFromConnectionString()` and the corresponding
test scenario.